renamed DomainEventContainer to DomainEventRepository
